refactor: split contract.rs
fix: allow clippt module inception

refactor: resplit based on components
