+++ src/integration_tests.rs
@@ -1,4 +1,5 @@
 use crate::helpers::CwTemplateContract;
+use crate::msg::ExecuteMsg;
 use crate::msg::InstantiateMsg;
 use cosmwasm_std::{Addr, Coin, Empty, Uint128};
 use cw_multi_test::{App, AppBuilder, Contract, ContractWrapper, Executor};
@@ -56,18 +57,13 @@ fn proper_instantiate() -> (App, CwTemplateContract) {
     (app, cw_template_contract)
 }
 
-mod data_requests {
-    use super::*;
-    use crate::msg::ExecuteMsg;
+#[test]
+fn post_data_request() {
+    let (mut app, cw_template_contract) = proper_instantiate();
 
-    #[test]
-    fn post_data_request() {
-        let (mut app, cw_template_contract) = proper_instantiate();
-
-        let msg = ExecuteMsg::PostDataRequest {
-            value: "hello world".to_string(),
-        };
-        let cosmos_msg = cw_template_contract.call(msg).unwrap();
-        app.execute(Addr::unchecked(USER), cosmos_msg).unwrap();
-    }
+    let msg = ExecuteMsg::PostDataRequest {
+        value: "hello world".to_string(),
+    };
+    let cosmos_msg = cw_template_contract.call(msg).unwrap();
+    app.execute(Addr::unchecked(USER), cosmos_msg).unwrap();
 }
