+++ src/lib.rs
@@ -1,9 +1,17 @@
 pub mod consts;
 pub mod contract;
+pub mod data_request;
+pub mod data_request_result;
 mod error;
+pub mod executors_registry;
 pub mod helpers;
 pub mod msg;
+pub mod staking;
 pub mod state;
-pub mod tests;
-
 pub use crate::error::ContractError;
+
+#[path = ""]
+#[cfg(test)]
+mod tests {
+    mod integration_tests;
+}
