--- src/tests.rs
@@ -1,5 +1,0 @@
-#[path = ""]
-#[cfg(test)]
-mod tests {
-    mod integration_tests;
-}
